Adição strategy

+++ LibraDesk/src/strategy/CalculadoraMulta.java
@@ -1,5 +1,5 @@
 package strategy;
 
 public interface CalculadoraMulta {
-    double calcularMulta(int diasAtraso);
+    double calcularMulta(long diasAtraso);
 }
