+++ LibraDesk/src/strategy/MultaEspecial.java
@@ -3,7 +3,7 @@ package strategy;
 public class MultaEspecial implements CalculadoraMulta {
     
     @Override
-    public double calcularMulta(int diasAtraso) {
+    public double calcularMulta(long diasAtraso) {
         return ((diasAtraso - 30) * 2) + 30;
     }
 }
