+++ LibraDesk/src/strategy/MultaPadrao.java
@@ -3,7 +3,7 @@ package strategy;
 public class MultaPadrao implements CalculadoraMulta {
     
     @Override
-    public double calcularMulta(int diasAtraso) {
+    public double calcularMulta(long diasAtraso) {
         return diasAtraso;
     }
 }
